修改了发布任务不能发布Double佣金的Bug 修改了退款的bug 修改了N个小Bug 家安 11.13

+++ app/src/main/java/com/power/mercenary/activity/PubGerendingzhiActivity.java
@@ -472,7 +472,7 @@ public class PubGerendingzhiActivity extends BaseActivity implements PubTaskPres
 
                 picNumsPresenter.publishTask(fileList);//上传多张图片
 
-                presenter.publishTask("", taskType, childTaskType, taskNameEt.getText().toString(), s1, imgurl, (Integer.parseInt(taskMoneyEt.getText().toString())*100)+"",
+                presenter.publishTask("", taskType, childTaskType, taskNameEt.getText().toString(), s1, imgurl, (Double.parseDouble(taskMoneyEt.getText().toString())*100)+"",
                         "", taskDesEt.getText().toString(), "", s,
                         "", "", "", "",
                         "", "", "");
