+++ app/src/main/java/com/power/mercenary/activity/PubQitaActivity.java
@@ -351,7 +351,7 @@ public class PubQitaActivity extends BaseActivity implements PubTaskPresenter.Pu
                 String s = MyUtils.listToString(requireList);
                 String s1 = MyUtils.listToString(biaoqianList);
 
-                presenter.publishTask("", taskType, childTaskType, taskNameEt.getText().toString(), s1, "", (Integer.parseInt(taskMoneyEt.getText().toString())*100)+"",
+                presenter.publishTask("", taskType, childTaskType, taskNameEt.getText().toString(), s1, "", (Double.parseDouble(taskMoneyEt.getText().toString())*100)+"",
                         validityTimeEt.getText().toString(), taskDetailEt.getText().toString(), taskMudiEt.getText().toString(), s,
                         "", "", "", "",
                         "", "", "");
